javadoc and refactoring
