+++ student-demo-api/src/main/java/nilotpal/jaxrs/init/MyApplicationConfig.java
@@ -8,6 +8,9 @@ import org.slf4j.LoggerFactory;
 
 import javax.ws.rs.ApplicationPath;
 
+/**
+ * Custom Resource configuration class to register the required resource classes
+ */
 @ApplicationPath("/api")
 public class MyApplicationConfig extends ResourceConfig {
     private final static Logger log = LoggerFactory.getLogger(MyApplicationConfig.class);
