--- student-demo-api/src/main/java/nilotpal/jaxrs/init/MyApplicationFeature.java
@@ -1,21 +1,0 @@
-package nilotpal.jaxrs.init;
-
-import org.glassfish.hk2.api.ServiceLocator;
-import org.glassfish.hk2.utilities.ServiceLocatorUtilities;
-
-import javax.inject.Inject;
-import javax.ws.rs.core.Feature;
-import javax.ws.rs.core.FeatureContext;
-
-public class MyApplicationFeature implements Feature {
-
-    @Inject
-    public MyApplicationFeature(ServiceLocator serviceLocator) {
-        ServiceLocatorUtilities.enableImmediateScope(serviceLocator);
-    }
-
-    @Override
-    public boolean configure(FeatureContext context) {
-        return true;
-    }
-}
